requirements.txt
Pillow==8.4.0
diffusers==0.14.0
torch
numpy
transformers
xformers
accelerate
imageio
flask
boto3